just need to read and write

+++ include/core/bus.h
@@ -8,10 +8,10 @@ struct Bus
 {
   Bus(bool skip, std::string bootrom_path);
   u8 ReadByte(u16 addr);
-  u8 NextByte(u16 addr, u16& pc, u64& cycles);
+  u8 NextByte(u16 addr, u16& pc, u8& cycles);
   void WriteByte(u16 addr, u8 val);
   u16 ReadHalf(u16 addr);
-  u16 NextHalf(u16 addr, u16& pc, u64& cycles);
+  u16 NextHalf(u16 addr, u16& pc, u8& cycles);
   void WriteHalf(u16 addr, u16 val);
   void LoadROM(std::string filename);
   void Reset();
